+++ scripts/notifier.py
@@ -1,10 +1,10 @@
 """
-This script is an example of 'exchangelib' usage. It will give you email and appointment notifications from your 
+This script is an example of 'exchangelib' usage. It will give you email and appointment notifications from your
 Exchange account on your Ubuntu desktop.
 
 Usage: notifier.py [notify_interval]
 
-You need to install the `libxml2-dev` `libxslt1-dev` packages for 
+You need to install the `libxml2-dev` `libxslt1-dev` packages for
 'exchangelib' to work on Ubuntu.
 
 Login and password is fetched from `~/.netrc`. Add an entry like this:
@@ -73,7 +73,7 @@ now = UTC_NOW()
 emails_since = now - timedelta(seconds=sleep)
 cal_items_before = now + timedelta(seconds=sleep * 4)  # Longer notice of upcoming appointments than new emails
 username, _, password = netrc().authenticators('office365')
-c = Credentials(username, password, is_service_account=False)
+c = Credentials(username, password)
 a = Account(primary_smtp_address=c.username, credentials=c, access_type=DELEGATE, autodiscover=True, verify_ssl=False)
 
 for msg in a.calendar.view(start=now, end=cal_items_before)\
@@ -84,9 +84,9 @@ for msg in a.calendar.view(start=now, end=cal_items_before)\
     minutes_to_appointment = int((msg.start - now).total_seconds() / 60)
     subj = 'You have a meeting in %s minutes' % minutes_to_appointment
     body = '%s-%s: %s\n%s' % (
-        msg.start.astimezone(tz).strftime('%H:%M'),  
-        msg.end.astimezone(tz).strftime('%H:%M'), 
-        msg.subject[:150], 
+        msg.start.astimezone(tz).strftime('%H:%M'),
+        msg.end.astimezone(tz).strftime('%H:%M'),
+        msg.subject[:150],
         msg.location
     )
     zenity(**{'info': None, 'no-markup': None, 'title': subj, 'text': body})